requirements.txt
asgiref==3.8.1
Django==5.1.6
django-autoslug==1.9.9
django-crispy-forms==2.3
pillow==11.1.0
PyMySQL==1.1.1
sqlparse==0.5.3
tzdata==2025.1
